refactor: Rename activity_document_mongo to activity_document
