+++ backend/src/adapters/output/activity_document.py
