+++ backend/src/adapters/output/activity_repository_firestore_adapter.py
@@ -3,7 +3,7 @@ import logging
 from uuid import UUID
 
 from google.cloud import firestore
-from src.adapters.output.activity_document_mongo import ActivityDocument
+from src.adapters.output.activity_document import ActivityDocument
 from src.application.ports.output.activity_repository import ActivityRepositoryPort
 from src.domain.entities.activity import Activity
