+++ backend/src/config/database.py
@@ -2,7 +2,7 @@ import logging
 
 from beanie import init_beanie
 from motor.motor_asyncio import AsyncIOMotorClient
-from src.adapters.output.activity_document_mongo import ActivityDocument
+from src.adapters.output.activity_document import ActivityDocument
 
 logger = logging.getLogger(__name__)
