+++ backend/tests/unit/adapters/output/test_activity_document.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING
 
 import pytest
-from src.adapters.output.activity_document_mongo import ActivityDocument
+from src.adapters.output.activity_document import ActivityDocument
 
 if TYPE_CHECKING:
     from src.domain.entities.activity import Activity
