+++ backend/tests/unit/adapters/output/test_activity_repository_mongo_adapter.py
@@ -2,7 +2,7 @@ from dataclasses import replace
 from unittest.mock import AsyncMock, patch
 
 import pytest
-from src.adapters.output.activity_document_mongo import ActivityDocument
+from src.adapters.output.activity_document import ActivityDocument
 from src.adapters.output.activity_repository_mongo_adapter import ActivityRepositoryAdapter
